updated decimal to float

+++ db/migrate/20220602104646_add_volume24_to_collections.rb
@@ -1,5 +1,5 @@
 class AddVolume24ToCollections < ActiveRecord::Migration[6.1]
   def change
-    add_column :collections, :volume24, :decimal
+    add_column :collections, :volume24, :float
   end
 end
